dom/system/tests/test-notaccessiblefromcontent.html

<!DOCTYPE HTML>
<html>
<head>
 <title>Testing that OS.Constants is not available from web content</title>
 <script src="/tests/SimpleTest/SimpleTest.js"></script>
 <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
 <script type="application/javascript">

// 1. Test that OS is not available from the main thread
 SimpleTest.is(typeof OS, "undefined", "OS is not available from the main thread");

// 2. Test that OS is not available from a web worker
  SimpleTest.waitForExplicitFinish();
  worker = new Worker("worker_notaccessiblefromcontent.js");
  worker.onmessage = function onmessage(event) {
   SimpleTest.info("Received " + JSON.stringify(event.data));
   SimpleTest.ok(event.data.result, "OS is not available from a web worker");
   SimpleTest.finish();
  }
  worker.onerror = function onerror(event) {
   SimpleTest.ok(false, "Error: " + event);
   SimpleTest.finish();
  };
  worker.postMessage(0);

</script>
</head>

<body>
</body>
</html>
